pyproject.toml
[project]
name = "talkback-api"
version = "1.0.0"
description = "Python Fast API for the TalkBack Starter App"
readme = "README.md"
license = {file = "LICENSE.md"}
dependencies = [
  "alembic",
  "boto3",
  "fastapi",
  "google-cloud-texttospeech",
  "httpx",
  "pydantic-settings",
  "python-dotenv",
  "python-jose[cryptography]",
  "requests",
  "sqlalchemy",
  "uvicorn[standard]",
]

[project.optional-dependencies]
dev = ["coverage", "pytest", "pytest-asyncio", "ruff"]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
exclude = [
    ".git",
    ".mypy_cache",
    ".pre-commit-cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "venv",
    "docs",
    "__pycache",
    "**/migrations/*",
]

line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["B", "C4", "EXE", "F", "E", "ISC", "ICN", "INP", "PIE", "SIM", "W", "T20", "UP", "T10", "G", "C90", "ERA"]
ignore = ["B008", "SIM102"]
fixable = ["F", "E", "B", "C4", "EXE", "ISC", "ICN", "INP", "PIE", "SIM", "W", "T20", "UP"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short"]
